feat(test): T08_S02 PTY 스트리밍 통합 테스트 구현
- 포괄적인 통합 테스트 프레임워크 구현
- PTY 세션 라이프사이클 테스트 완료
- WebSocket 스트리밍 및 다중 클라이언트 테스트
- Docker 컨테이너 통합 테스트 (재시작, 여러 이미지 타입)
- 성능 테스트 (동시 세션, 메모리 사용량)
- 테스트 클라이언트 및 Docker 관리자 구현
- 메트릭 수집 및 리포트 생성 기능

컴파일 오류 수정:
- gzip.SetConcurrency 제거 (표준 라이브러리에 없음)
- 사용하지 않는 import 제거
- 중복 함수 선언 해결 (generateSessionID)
- log 변수 충돌 해결
- 타입 불일치 수정 (uint32/uint64)
- import 경로 수정 (github.com/aicli-web -> github.com/aicli/aicli-web)

+++ internal/flow/throttler.go
@@ -525,7 +525,7 @@ func calculatePerformance(state *ThrottleState) float64 {
 	}
 	
 	// 스로틀링 비율의 역수를 성능 지표로 사용
-	totalMessages := float64(state.ThrottledCount + state.AdjustmentCount*100)
+	totalMessages := float64(state.ThrottledCount + uint64(state.AdjustmentCount)*100)
 	if totalMessages == 0 {
 		return 1.0
 	}
